requirements.txt
youtube-transcript-api 
langchain-community 
langchain-openai 
langchain-huggingface 
faiss-cpu 
tiktoken 
python-dotenv
langchain
langchain-text-splitters
sentence-transformers
Flask
flask-cors
python-dotenv
